Update gui
Remove automatic plotting

Still errors with calling previous config files

Currently code is commented out instead of deleted.
